Event context
Created event context. Changed the Association provider. Changed imports. Removed context directory. Updated imports of the "context" directory to be "contexts"
